floats

+++ pytorch/data.py
@@ -30,8 +30,8 @@ class LabeledKaggleDataset(data.Dataset):
         f = self.lung_names[index] + '.npy'
         img = load_img(self.image_dir + f)
         img = hu_to_visual_features(img, -1500, 500)
-        img = torch.from_numpy(img)
-        target = torch.DoubleTensor(1, 1)
+        img = torch.from_numpy(img).float()
+        target = torch.FloatTensor(1, 1)
         target[0,0] = self.lung_labels[index]
         if self.input_transform:
             img = self.input_transform(img)
